LayoutTests/platform/efl/accessibility/img-fallsback-to-title-expected.txt

     test
test
This tests that images will fallback to using the title attribute if no other descriptive text is present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Image1 description: AXDescription: test1
Image1 help: null

Image2 description: AXDescription: alt
Image2 help: null

Image3 description: AXDescription: test3
Image3 help: null

Image4 description: AXDescription: alt
Image4 help: null

PASS imagesGroup.childAtIndex(0).childrenCount is 2
PASS successfullyParsed is true

TEST COMPLETE
